chore: create helper to handle Skip logic and SafeArea inset
