+++ cardstack/src/screens/Profile/ProfileSlugScreen/useProfileSlugScreen.ts
@@ -1,11 +1,9 @@
 import { validateMerchantId } from '@cardstack/cardpay-sdk';
-import { StackActions, useNavigation } from '@react-navigation/native';
+import { useNavigation } from '@react-navigation/native';
 import { useState, useCallback, useEffect, useMemo, useRef } from 'react';
-import { useDispatch } from 'react-redux';
 
 import { useInitIAPProducts } from '@cardstack/hooks/usePurchaseProfile';
 import { Routes } from '@cardstack/navigation';
-import { skipProfileCreation } from '@cardstack/redux/persistedFlagsSlice';
 import { useLazyValidateProfileSlugQuery } from '@cardstack/services';
 import { matchMinLength } from '@cardstack/utils/validators';
 
@@ -14,8 +12,7 @@ import { strings, MIN_SLUG_LENGTH } from './strings';
 export const useProfileSlugScreen = () => {
   useInitIAPProducts();
 
-  const { navigate, dispatch: navDispatch } = useNavigation();
-  const dispatch = useDispatch();
+  const { navigate } = useNavigation();
 
   const [slug, setSlug] = useState('');
 
@@ -82,17 +79,11 @@ export const useProfileSlugScreen = () => {
     [isFetching, isLoading, slugValidation.slugAvailable]
   );
 
-  const onSkipPress = useCallback(() => {
-    dispatch(skipProfileCreation(true));
-    navDispatch(StackActions.pop(1));
-  }, [navDispatch, dispatch]);
-
   return {
     slug,
     onSlugChange,
     onContinuePress,
     slugValidation,
     canContinue,
-    onSkipPress,
   };
 };
